requirements.txt
feedparser>=6.0.0
requests>=2.31.0
python-dateutil>=2.8.0
anthropic>=0.40.0
python-dotenv>=1.0.0
schedule>=1.2.0

